schema/context: restore some backlinks support
In libyang v1 the schema nodes had a backlinks member to be able to
look up dependents of the node.  SONiC depends on this to provide
functionality it uses and it needs to be exposed via the python
module.

In theory, exposing the 'dfs' functions could make this work, but
it would likely be cost prohibitive since walking the tree would
be expensive to create a python node for evaluation in native
python.

This implementation provides 2 python functions:
 * Context.backlinks_find_leafref_nodes(path) - This function can
   take the path of the base node and find all dependents.  If
   no path is specified, then it will return all nodes that contain
   a leafref reference.
 * Context.backlinks_xpath_leafrefs(xpath) - This function takes
   an xpath, then returns all target nodes the xpath may reference.
   Typically only one will be returned, but multiples may be in the
   case of a union.

A user can build a cache by combining Context.backlinks_find_leafref_nodes()
with no path set and building a reverse table using
Context.backlinks_xpath_leafrefs(xpath)

Signed-off-by: Brad House <brad@brad-house.com>